fix: some packages not being downloaded (#108)
* feat: use Environment.NewLine instead of \n

* feat: add a provider that provides nuget caches

* feat: add a component that can find packages artifact dirs

* feat: prepare for artifact dir enrichment

* refactor: remove unnecessary GetDownloadedPackageReferenceArtifact

* feat: actually enrich with artifact directories

* chore: exclude optional from code coverage

* test: add missing test for deleting temp directory

* feat: download packages where we found no artifact directory

* feat: show target-file name when settings invalid because of target-file

* feat: remove project-reference package-references

* test: add missing test in GetPackageReferencesViaDotnetCli

* refactor: remove unnecessary nullable string from settings

* refactor: fix some code issues

+++ src/Core/Liz.Core/Extract/ExtractLicenses.cs
@@ -21,9 +21,10 @@ namespace Liz.Core.Extract;
 internal sealed class ExtractLicenses : IExtractLicenses
 {
     private readonly IGetPackageReferences _getPackageReferences;
+    private readonly IEnrichPackageReferenceWithArtifactDirectory _enrichPackageReferenceWithArtifactDirectory;
+    private readonly IDownloadPackageReferences _downloadPackageReferences;
     private readonly IEnrichPackageReferenceWithLicenseInformation _enrichPackageReferenceWithLicenseInformation;
     private readonly IProvideTemporaryDirectories _provideTemporaryDirectories;
-    private readonly IDownloadPackageReferences _downloadPackageReferences;
     private readonly IEnumerable<IResultProcessor> _resultProcessors;
     private readonly IEnumerable<IPreprocessor> _preprocessors;
     private readonly IGetProjects _getProjects;
@@ -37,20 +38,24 @@ internal sealed class ExtractLicenses : IExtractLicenses
         IProgressHandler? progressHandler,
         IGetProjects getProjects,
         IGetPackageReferences getPackageReferences,
+        IEnrichPackageReferenceWithArtifactDirectory enrichPackageReferenceWithArtifactDirectory,
+        IDownloadPackageReferences downloadPackageReferences,
         IEnrichPackageReferenceWithLicenseInformation enrichPackageReferenceWithLicenseInformation,
         IProvideTemporaryDirectories provideTemporaryDirectories,
-        IDownloadPackageReferences downloadPackageReferences,
         IEnumerable<IResultProcessor> resultProcessors,
         IEnumerable<IPreprocessor> preprocessors)
     {
         _settings = settings ?? throw new ArgumentNullException(nameof(settings));
         _getProjects = getProjects ?? throw new ArgumentNullException(nameof(getProjects));
         _getPackageReferences = getPackageReferences ?? throw new ArgumentNullException(nameof(getPackageReferences));
+        _enrichPackageReferenceWithArtifactDirectory = enrichPackageReferenceWithArtifactDirectory ??
+                                                       throw new ArgumentNullException(
+                                                           nameof(enrichPackageReferenceWithArtifactDirectory));
+        _downloadPackageReferences = downloadPackageReferences ?? throw new ArgumentNullException(nameof(downloadPackageReferences));
         _enrichPackageReferenceWithLicenseInformation = enrichPackageReferenceWithLicenseInformation ?? 
                                                         throw new ArgumentNullException(nameof(enrichPackageReferenceWithLicenseInformation));
         _provideTemporaryDirectories = provideTemporaryDirectories 
                                      ?? throw new ArgumentNullException(nameof(provideTemporaryDirectories));
-        _downloadPackageReferences = downloadPackageReferences ?? throw new ArgumentNullException(nameof(downloadPackageReferences));
         _resultProcessors = resultProcessors ?? throw new ArgumentNullException(nameof(resultProcessors));
         _preprocessors = preprocessors ?? throw new ArgumentNullException(nameof(preprocessors));
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
@@ -62,22 +67,15 @@ internal sealed class ExtractLicenses : IExtractLicenses
         try
         {
             await PrepareAsync().ConfigureAwait(false);
-            var projects = GetProjects(_settings.GetTargetFile()!).ToList();
-
-            await DownloadPackageReferencesFor(projects).ConfigureAwait(false);
+            var projects = GetProjects(_settings.GetTargetFile()).ToList();
             
             var packageReferences = (await GetPackageReferencesAsync(projects).ConfigureAwait(false)).ToList();
+
+            await EnrichWithArtifactDirectory(packageReferences).ConfigureAwait(false);
+            await DownloadPackagesWithoutArtifactDirectory(packageReferences).ConfigureAwait(false);
+            
             await EnrichWithLicenseInformationAsync(packageReferences).ConfigureAwait(false);
 
-            packageReferences = packageReferences
-                /*
-                 * this gets rid of all the references that have NO data in the license-information what so ever
-                 * (which basically are the internal project-references or other packages that have no information or
-                 * were not downloaded)
-                 */
-                .Where(packageReference => !packageReference.LicenseInformation.IsEmpty())
-                .ToList();
-            
             _progressHandler?.FinishMainProcess();
 
             await ProcessResultsAsync(packageReferences).ConfigureAwait(false);
@@ -99,14 +97,15 @@ internal sealed class ExtractLicenses : IExtractLicenses
         try
         {
             /*
-             * 5 main process steps:
+             * 6 main process steps:
              * - preparation
              * - get projects
-             * - download package-references
              * - get package-references
-             * - enrich license info
+             * - enrich with artifact-directory
+             * - download packages without artifact-directory
+             * - enrich with license info
              */
-            _progressHandler?.InitializeMainProcess(5, "Preparing");
+            _progressHandler?.InitializeMainProcess(6, "Preparing");
             _logger.LogInformation("Preparing...");
 
             foreach (var preprocessor in _preprocessors)
@@ -139,34 +138,16 @@ internal sealed class ExtractLicenses : IExtractLicenses
         }
     }
 
-    private async Task DownloadPackageReferencesFor(IEnumerable<Project> projects)
-    {
-        var projectsList = projects.ToList();
-        
-        _logger.LogInformation("Downloading package-references for project(s)...");
-     
-        _progressHandler?.TickMainProcess("Download package-references");
-        _progressHandler?.InitializeNewSubProcess(projectsList.Count);
-        
-        foreach (var project in projectsList)
-        {
-            _progressHandler?.TickCurrentSubProcess($"Download package-references: {project.Name}");
-            
-            await _downloadPackageReferences.DownloadForProjectAsync(project).ConfigureAwait(false);
-        }
-    }
-
-    private async Task<IEnumerable<PackageReference>> GetPackageReferencesAsync(IEnumerable<Project> projects)
+    private async Task<IEnumerable<PackageReference>> GetPackageReferencesAsync(IReadOnlyCollection<Project> projects)
     {
         var packageReferences = new List<PackageReference>();
-        var projectsList = projects.ToList();
-        
+
         _logger.LogInformation("Trying to get package-references for project(s)...");
 
         _progressHandler?.TickMainProcess("Get package-references");
-        _progressHandler?.InitializeNewSubProcess(projectsList.Count);
+        _progressHandler?.InitializeNewSubProcess(projects.Count);
         
-        foreach (var project in projectsList)
+        foreach (var project in projects)
         {
             _progressHandler?.TickCurrentSubProcess($"Get package-references: {project.Name}");
 
@@ -179,6 +160,50 @@ internal sealed class ExtractLicenses : IExtractLicenses
             .ToList();
         
         return packageReferences;
+    }
+
+    private async Task EnrichWithArtifactDirectory(IReadOnlyCollection<PackageReference> packageReferences)
+    {
+        _logger.LogInformation("Trying to get artifacts for package(s)...");
+        
+        _progressHandler?.TickMainProcess("Get artifacts");
+        _progressHandler?.InitializeNewSubProcess(packageReferences.Count);
+
+        foreach (var packageReference in packageReferences)
+        {
+            _progressHandler?.TickCurrentSubProcess($"Get artifact: {packageReference.Name}");
+            await EnrichWithArtifactDirectoryForPackageReferenceAsync(packageReference).ConfigureAwait(false);
+        }
+    }
+
+    private async Task DownloadPackagesWithoutArtifactDirectory(IEnumerable<PackageReference> packageReferences)
+    {
+        _logger.LogInformation("Downloading packages that are not in the cache...");
+
+        _progressHandler?.TickMainProcess("Download packages");
+        // we'll download everything in one go
+        _progressHandler?.InitializeNewSubProcess(1);
+
+        var packagesWithoutArtifactDirectory = packageReferences
+            .Where(package => package.ArtifactDirectory == null);
+
+        await _downloadPackageReferences.DownloadAndEnrichAsync(packagesWithoutArtifactDirectory).ConfigureAwait(false);
+    }
+
+    private async Task EnrichWithArtifactDirectoryForPackageReferenceAsync(PackageReference packageReference)
+    {
+        try
+        {
+            _logger.LogDebug($"Trying to get artifact directory for {packageReference}...");
+
+            await _enrichPackageReferenceWithArtifactDirectory.EnrichAsync(packageReference).ConfigureAwait(false);
+            
+            _logger.LogDebug($"Found following artifact directory: '{packageReference.ArtifactDirectory}'");
+        }
+        catch (Exception exception)
+        {
+            throw new GetArtifactDirectoryFailedException(packageReference, exception);
+        }
     }
 
     private async Task<IEnumerable<PackageReference>> GetPackageReferencesForProjectAsync(Project project)
@@ -204,16 +229,14 @@ internal sealed class ExtractLicenses : IExtractLicenses
         }
     }
 
-    private async Task EnrichWithLicenseInformationAsync(IEnumerable<PackageReference> packageReferences)
+    private async Task EnrichWithLicenseInformationAsync(IReadOnlyCollection<PackageReference> packageReferences)
     {
-        var packageReferencesList = packageReferences.ToList();
-        
         _logger.LogInformation("Trying to get license information for package(s)...");
         
         _progressHandler?.TickMainProcess("Get license information");
-        _progressHandler?.InitializeNewSubProcess(packageReferencesList.Count);
+        _progressHandler?.InitializeNewSubProcess(packageReferences.Count);
 
-        foreach (var packageReference in packageReferencesList)
+        foreach (var packageReference in packageReferences)
         {
             _progressHandler?.TickCurrentSubProcess($"Get license information: {packageReference.Name}");
 
@@ -251,6 +274,8 @@ internal sealed class ExtractLicenses : IExtractLicenses
     private void CleanUpTemporaryDirectory()
     {
         var temporaryDirectory = _provideTemporaryDirectories.GetRootDirectory();
+
+        if (!temporaryDirectory.Exists) return;
         temporaryDirectory.Delete(true);
     }
 }
